refactor: sanitize code
