+++ src/control/states/AnimatingState.ts
@@ -5,8 +5,7 @@
 import { ComponentEvent } from "@egjs/component";
 
 import { EVENTS } from "../../const/external";
-import * as AXES from "../../const/axes";
-import { circulatePosition, getDirection } from "../../utils";
+import { getDirection } from "../../utils";
 
 import State, { STATE_TYPE } from "./State";
 
@@ -46,35 +45,7 @@ class AnimatingState extends State {
   }
 
   public onChange(ctx: Parameters<State["onChange"]>[0]): void {
-    const { flicking, axesEvent, transitTo } = ctx;
-
-    if (!axesEvent.delta.flick) {
-      return;
-    }
-
-    const camera = flicking.camera;
-    const prevPosition = camera.position;
-    const position = axesEvent.pos[AXES.POSITION_KEY];
-    const newPosition = flicking.circularEnabled
-      ? circulatePosition(position, camera.range.min, camera.range.max)
-      : position;
-
-    void camera.lookAt(newPosition);
-
-    const moveEvent = new ComponentEvent(EVENTS.MOVE, {
-      isTrusted: axesEvent.isTrusted,
-      holding: this.holding,
-      direction: getDirection(0, axesEvent.delta.flick),
-      axesEvent
-    });
-
-    flicking.trigger(moveEvent);
-
-    if (moveEvent.isCanceled()) {
-      // Return to previous position
-      void flicking.camera.lookAt(prevPosition);
-      transitTo(STATE_TYPE.DISABLED);
-    }
+    this._moveToChangedPosition(ctx);
   }
 
   public onFinish(ctx: Parameters<State["onFinish"]>[0]) {
