+++ src/control/states/State.ts
@@ -4,8 +4,12 @@
  */
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { OnAnimationEnd, OnChange, OnFinish, OnHold, OnRelease } from "@egjs/axes";
+import { ComponentEvent } from "@egjs/component";
 
 import Flicking from "../../Flicking";
+import { EVENTS } from "../../const/external";
+import * as AXES from "../../const/axes";
+import { circulatePosition, getDirection } from "../../utils";
 
 export enum STATE_TYPE {
   IDLE,
@@ -123,6 +127,38 @@ abstract class State {
     transitTo: (nextState: STATE_TYPE) => State;
   }): void {
     // DO NOTHING
+  }
+
+  protected _moveToChangedPosition(ctx: Parameters<State["onChange"]>[0]): void {
+    const { flicking, axesEvent, transitTo } = ctx;
+
+    if (!axesEvent.delta[AXES.POSITION_KEY]) {
+      return;
+    }
+
+    const camera = flicking.camera;
+    const prevPosition = camera.position;
+    const position = axesEvent.pos[AXES.POSITION_KEY];
+    const newPosition = flicking.circularEnabled
+      ? circulatePosition(position, camera.range.min, camera.range.max)
+      : position;
+
+    void camera.lookAt(newPosition);
+
+    const moveEvent = new ComponentEvent(EVENTS.MOVE, {
+      isTrusted: axesEvent.isTrusted,
+      holding: this.holding,
+      direction: getDirection(0, axesEvent.delta[AXES.POSITION_KEY]),
+      axesEvent
+    });
+
+    flicking.trigger(moveEvent);
+
+    if (moveEvent.isCanceled()) {
+      // Return to previous position
+      void camera.lookAt(prevPosition);
+      transitTo(STATE_TYPE.DISABLED);
+    }
   }
 }
